M3/lesson.16_DRF_знакомство/pyproject.toml
[project]
name = "javarush-django-pyw"
version = "0.1.0"
description = ""
authors = [
    {name = "Stanislav",email = "stanislav3@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12, <4.0"
dependencies = [
    "django (>=5.2.5,<6.0.0)",
    "faker (>=37.8.0,<38.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)",
    "pytest-mock (>=3.15.1,<4.0.0)",
    "beautifulsoup4 (>=4.13.5,<5.0.0)",
    "celery[redis] (>=5.5.3,<6.0.0)",
    "django-celery-results (>=2.6.0,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "django-celery-beat (>=2.8.1,<3.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "requests (>=2.32.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
